+++ lib/models/apphud_models/apphud_subscription.dart
@@ -12,13 +12,7 @@ class ApphudSubscriptionWrapper {
   final bool isActive;
   final ApphudSubscriptionStatus status;
   final double? canceledAt;
-
-  // TODO: startedAt не может быть null, ошибка в Android либе. можно смело делать required
-  final double? startedAt;
-
-  // TODO: remove isSandbox & isLocal, because they are not present in Android
-  final bool? isSandbox;
-  final bool? isLocal;
+  final double startedAt;
 
   ApphudSubscriptionWrapper({
     required this.productId,
@@ -28,10 +22,8 @@ class ApphudSubscriptionWrapper {
     required this.isIntroductoryActivated,
     required this.isActive,
     required this.status,
-    this.startedAt,
+    required this.startedAt,
     this.canceledAt,
-    this.isSandbox,
-    this.isLocal,
   });
 
   factory ApphudSubscriptionWrapper.fromJson(Map<dynamic, dynamic> map) =>
@@ -41,7 +33,7 @@ class ApphudSubscriptionWrapper {
 
   @override
   String toString() {
-    return 'ApphudSubscriptionWrapper{productId: $productId, expiresAt: $expiresAt, isInRetryBilling: $isInRetryBilling, isAutorenewEnabled: $isAutorenewEnabled, isIntroductoryActivated: $isIntroductoryActivated, isActive: $isActive, status: $status, canceledAt: $canceledAt, startedAt: $startedAt, isSandbox: $isSandbox, isLocal: $isLocal}';
+    return 'ApphudSubscriptionWrapper{productId: $productId, expiresAt: $expiresAt, isInRetryBilling: $isInRetryBilling, isAutorenewEnabled: $isAutorenewEnabled, isIntroductoryActivated: $isIntroductoryActivated, isActive: $isActive, status: $status, canceledAt: $canceledAt, startedAt: $startedAt}';
   }
 }