+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 - [Android] **BREAKING** Class 'AndroidPurchaseWrapper' was modified.
 - [Android], [iOS] **BREAKING** Class 'ApphudNonRenewingPurchase' was modified.
 - [Android], [iOS] **BREAKING** Class 'ApphudPaywall' was modified.
-
+- [Android], [iOS] **BREAKING** Class 'ApphudProduct' was modified.
+- [Android], [iOS] **BREAKING** Class 'ApphudSubscriptionWrapper' was modified.
 
 ## 2.3.2
 - [Android] Fixed restore purchase issue in some cases